new version of restbase and removed annoying "file not found" error for fix-symlinks.sh

+++ fix-symlinks.sh
@@ -8,10 +8,11 @@ CWD="$(
 )"
 
 #remove the existing symlinks
-rm $CWD/mediawiki/LocalSettings.php
-rm $CWD/mediawiki/skins/Neverland
+test -f $CWD/mediawiki/LocalSettings.php && rm $CWD/mediawiki/LocalSettings.php
+test -f $CWD/mediawiki/skins/Neverland && rm $CWD/mediawiki/skins/Neverland
+test -f $CWD/mediawiki/favicon.ico && rm $CWD/mediawiki/favicon.ico
 rm -r $CWD/mediawiki/extensions
-rm $CWD/mediawiki/favicon.ico
+
 
 #create symlinks
 ln -s ../LocalSettings.php mediawiki/LocalSettings.php
